requirement.txt


# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file> 
# platform: linux-64
# This two first libraries are to be installed using:
# $ pip3 install <library>

#pyyaml=6.0=pypi_0 
#pyaml==21.10.1=pypi_0

_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_py-xgboost-mutex=2.0=gpu_0
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.3=py38h0a891b7_0
aiosignal=1.2.0=pyhd8ed1ab_0
anyio=3.6.2=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h0a891b7_2
asttokens=2.0.8=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
babel=2.10.3=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
blinker=1.5=pyhd8ed1ab_0
bottleneck=1.3.5=py38h71d37f0_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py38h0a891b7_1004
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.9.24=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pyhd8ed1ab_0
cartopy=0.21.0=py38h606536b_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.14.4=py38ha312104_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py38h578d9bd_0
cloudpickle=2.2.0=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
contourpy=1.0.5=py38h43d8883_0
cryptography=38.0.2=py38h2b5fc30_0
cudatoolkit=11.7.0=hd8887f6_10
cudnn=8.4.1.50=hed8a83a_0
cycler=0.11.0=pyhd8ed1ab_0
deap=1.3.3=py38h47df419_0
debugpy=1.6.3=py38hfa26641_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.1.1=pyhd8ed1ab_0
flatbuffers=2.0.7=h27087fc_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.38.0=py38h0a891b7_0
freetype=2.12.1=hca18f0e_0
frozenlist=1.3.1=py38h0a891b7_0
gast=0.4.0=pyh9f0ad1d_0
geos=3.11.0=h27087fc_0
giflib=5.2.1=h36c2ea0_2
google-auth=2.13.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.47.1=h05bd8bd_6
grpcio=1.47.1=py38h7dc2bf5_6
h5py=3.7.0=nompi_py38h045baee_101
hdf5=1.12.2=nompi_h2386368_100
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py38h578d9bd_0
importlib_resources=5.10.0=pyhd8ed1ab_0
ipykernel=6.16.1=pyh210e3f2_0
ipython=8.5.0=pyh41d4057_1
ipython_genutils=0.2.0=py_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter_client=7.4.3=pyhd8ed1ab_0
jupyter_core=4.11.1=py38h578d9bd_0
jupyter_server=1.21.0=pyhd8ed1ab_0
jupyterlab=3.4.8=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.16.1=pyhd8ed1ab_0
keras=2.10.0=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h43d8883_0
krb5=1.19.3=h3790be6_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.39=hc81fddc_0
lerc=4.0.0=h27087fc_0
libabseil=20220623.0=cxx17_h48a1fff_4
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=16_linux64_openblas
libcurl=7.85.0=h7bff187_0
libdeflate=1.14=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=12.2.0=h65d4601_18
libgfortran-ng=12.2.0=h69a702a_18
libgfortran5=12.2.0=h337968e_18
libgomp=12.2.0=h65d4601_18
liblapack=3.9.0=16_linux64_openblas
libllvm11=11.1.0=he0ac6c6_5
libnghttp2=1.47.0=hdcd2b5c_1
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.38=h753d276_0
libprotobuf=3.21.8=h6239696_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.39.4=h753d276_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=12.2.0=h46fd767_18
libtiff=4.4.0=h55922b4_4
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxgboost=1.5.1=cuda112h47702df_2
libzlib=1.2.13=h166bdaf_4
llvmlite=0.39.1=py38h38d86a4_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_1
matplotlib-base=3.6.1=py38hb021067_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
multidict=6.0.2=py38h0a891b7_1
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.4.7=pyhd8ed1ab_0
nbclient=0.7.0=pyhd8ed1ab_0
nbconvert=7.2.2=pyhd8ed1ab_0
nbconvert-core=7.2.2=pyhd8ed1ab_0
nbconvert-pandoc=7.2.2=pyhd8ed1ab_0
nbformat=5.7.0=pyhd8ed1ab_0
nccl=2.14.3.1=h0800d71_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
notebook=6.4.12=pyha770c72_0
notebook-shim=0.2.0=pyhd8ed1ab_0
numba=0.56.4=py38h9a4aae9_0
numexpr=2.8.3=py38h97b1c41_100
numpy=1.23.4=py38h7042d01_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=h7d73246_1
openssl=1.1.1s=h166bdaf_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.0=py38h43a58ef_0
pandoc=2.19.2=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py38ha3b2c9c_2
pip=22.3=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
proj=9.1.0=h93bde94_0
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
protobuf=4.21.8=py38hfa26641_0
psutil=5.9.3=py38h0a891b7_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-xgboost=1.5.1=cuda112py38hded6998_2
#pyaml=21.10.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyproj=3.4.0=py38hd7890fc_1
pyrsistent=0.18.1=py38h0a891b7_1
pyshp=2.3.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.0=h357f687_5
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-flatbuffers=2.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2022.5=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=24.0.1=py38hfc09fa9_0
re2=2022.06.01=h27087fc_0
readline=8.1.2=h0f457ee_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.2=py38h0b08f9b_0
scipy=1.9.3=py38h8ce737c_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.5.0=pyhd8ed1ab_0
shap=0.41.0=py38h47df419_0
shapely=1.8.5=py38hc9bb657_0
six=1.16.0=pyh6c4a22f_0
slicer=0.0.7=pyhd8ed1ab_0
snappy=1.1.9=hbd366e4_1
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.39.4=h4ff8645_0
stack_data=0.5.1=pyhd8ed1ab_0
stopit=1.1.2=py_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38h2b5fc30_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.10.0=cuda112py38hded6998_0
tensorflow-base=2.10.0=cuda112py38h6b2b66c_0
tensorflow-estimator=2.10.0=cuda112py38hf5dcc89_0
termcolor=2.0.1=pyhd8ed1ab_1
terminado=0.16.0=pyh41d4057_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.2=py38h0a891b7_0
tpot=0.11.1=py_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.5.0=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
unicodedata2=14.0.0=py38h0a891b7_1
update_checker=0.18.0=pyh9f0ad1d_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.4.1=pyhd8ed1ab_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
wrapt=1.14.1=py38h0a891b7_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yarl=1.7.2=py38h0a891b7_2
zeromq=4.3.4=h9c3ff4c_1
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h6239696_4
